moved label to separate package
