+++ pkg/model/label/label.go
@@ -1,0 +1,7 @@
+package label
+
+func New(value string) string {
+	registry := GetRegistry()
+	registry.mustRegister(value)
+	return value
+}
