+++ pkg/model/model/globals.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"encoding/gob"
+	"github.com/praetorian-inc/tabularium/pkg/model/label"
 	"time"
 
 	"github.com/praetorian-inc/tabularium/pkg/registry"
@@ -106,7 +107,7 @@ type Insertable interface {
 }
 
 var (
-	TTLLabel = NewLabel("TTL") // TTLLabel attach to models that utilize TTL cleanup in neo4j
+	TTLLabel = label.New("TTL") // TTLLabel attach to models that utilize TTL cleanup in neo4j
 )
 
 const (
