+++ pkg/model/model/repository.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"fmt"
+	"github.com/praetorian-inc/tabularium/pkg/model/label"
 	"regexp"
 	"strings"
 
@@ -20,7 +21,7 @@ type Repository struct {
 }
 
 var (
-	RepositoryLabel = NewLabel("Repository")
+	RepositoryLabel = label.New("Repository")
 )
 
 var (
