+++ pkg/model/model/vulnerability.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"fmt"
+	"github.com/praetorian-inc/tabularium/pkg/model/label"
 
 	"github.com/praetorian-inc/tabularium/pkg/registry"
 )
@@ -31,7 +32,7 @@ func init() {
 	registry.Registry.MustRegisterModel(&VulnerabilityDefinition{})
 }
 
-var VulnerabilityLabel = NewLabel("Vulnerability")
+var VulnerabilityLabel = label.New("Vulnerability")
 
 func (v *Vulnerability) GetLabels() []string {
 	return []string{VulnerabilityLabel}
